implement order flow and add some changes

+++ src/main/java/com/khrystoforov/onlinestore/product/dto/request/ProductCreateRequestDto.java
@@ -13,5 +13,5 @@ public class ProductCreateRequestDto {
     private String name;
     private String description;
     private BigDecimal price;
-    private Long stockQuantity;
+    private Integer stockQuantity;
 }
